// src/types/partials/EncounterHospitalization.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TReference } from '../partials/Reference';
import { TCodeableConcept } from '../partials/CodeableConcept';

export type TEncounterHospitalization = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    preAdmissionIdentifier?: TIdentifier;
    origin?: TReference;
    admitSource?: TCodeableConcept;
    reAdmission?: TCodeableConcept;
    dietPreference?: TCodeableConcept[];
    specialCourtesy?: TCodeableConcept[];
    specialArrangement?: TCodeableConcept[];
    destination?: TReference;
    dischargeDisposition?: TCodeableConcept;
};